requirements.txt
pandas
pyyaml
pydub
soundfile
speechbrain
torch
torchvision
transformers==4.51.3
tqdm
peft
backoff
